add homework Four version 3
