+++ src/com/phonebook/Act.java
@@ -1,7 +1,5 @@
 package com.phonebook;
 
-import java.util.List;
-
 public abstract class Act {
     private final String name;
     private final int telephone;
@@ -9,10 +7,6 @@ public abstract class Act {
     public Act(String name, int telephone) {
         this.name = name;
         this.telephone = telephone;
-    }
-
-    public static void print(List<Record> list) {
-        list.forEach(System.out::println);
     }
 
     public String getName() {
